+++ src/main/java/com/bookingbustickets/bookingbustickets/domain/service/ScheduleService.java
@@ -7,7 +7,6 @@ import com.bookingbustickets.bookingbustickets.domain.repository.RouteRepository
 import com.bookingbustickets.bookingbustickets.domain.repository.ScheduleRepository;
 import com.bookingbustickets.bookingbustickets.exception.ScheduleNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -38,7 +37,7 @@ public class ScheduleService {
         if (optionalRoute.isEmpty()) {
             throw new ScheduleNotFoundException("Route with the given ID is not found");
         }
-        Schedule createdSchedule = new Schedule(requestScheduleDto.getScheduleDate(), requestScheduleDto.getDepartureTime(), requestScheduleDto.getArrivalTime(),optionalRoute.get());
+        Schedule createdSchedule = new Schedule(requestScheduleDto.getScheduleDate(), requestScheduleDto.getDepartureTime(), requestScheduleDto.getArrivalTime(), optionalRoute.get());
         return scheduleRepository.save(createdSchedule);
     }
 
@@ -51,9 +50,9 @@ public class ScheduleService {
     }
 
     public void deleteSchedule(Long id) {
-        try {
+        if (scheduleRepository.existsById(id)) {
             scheduleRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
+        } else {
             throw new ScheduleNotFoundException("Schedule with ID " + id + " is not found");
         }
     }
